Animation of the abundance spectra

Animation of the abundance spectra

animateSpectra(
  sim,
  species,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  power = 1,
  total = FALSE,
  resource = TRUE
)

Arguments

sim
A MizerSim object

species
Name or vector of names of the species to be plotted. By default all species are plotted.

wlim
A numeric vector of length two providing lower and upper limits for the w axis. Use NA to refer to the existing minimum or maximum.

ylim
A numeric vector of length two providing lower and upper limits for the y axis. Use NA to refer to the existing minimum or maximum. Any values below 1e-20 are always cut off.

power
The abundance is plotted as the number density times the weight raised to power. The default power = 1 gives the biomass density, whereas power = 2 gives the biomass density with respect to logarithmic size bins.

total
A boolean value that determines whether the total over all species in the system is plotted as well. Default is FALSE

resource
A boolean value that determines whether resource is included. Default is TRUE.

See also

Examples

# \dontrun{
sim <- project(NS_params, t_max = 15, effort = 1, progress_bar = FALSE)
animateSpectra(sim)

#> Warning: `arrange_()` is deprecated as of dplyr 0.7.0.
#> Please use `arrange()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

# }