Calculate a yield-vs-F curve

Calculates the yield of a species for a range of fishing mortalities for that species while the fishing mortalities for the other species are held fixed.

Usage

getYieldVsF(
  params,
  species,
  F_range,
  no_steps = 10,
  F_max = 1,
  F_min = 0,
  effort_it = 1,
  distance_func = distanceSSLogN,
  tol = 0.001,
  max_func = maxFthreshold,
  threshold_var = 10,
  t_max = 100
)

Arguments

params

An object of class MizerParams.

species

Name of the target species

F_range

A sequence of fishing mortalities at which to evaluate the yield. If missing, it is set to seq(from = 0, to = F_max, length.out = no_steps).

no_steps

The number of steps to use. Only used if F_range is missing.

F_max

The maximum fishing mortality. Used only if F_range is missing.

effort_it

The width of steps used to find maximum fisheries mortality, higher values means a faster calculation but less accurate maximum.

distance_func

A function that will be called after every t_per years with both the previous and the new state and that should return a number that in some sense measures the distance between the states. By default this uses the function distanceSSLogN() that you can use as a model for your own distance function.

tol

The projectToSteady function stops when the relative change in the egg production RDI over t_per years is less than tol for every species.

max_func

A function used in getMaxF to determines the threshold to use when stopping the fisheries effort.

threshold_var

A numeric value used in max_func to calculate the threshold from the params object. If supplied, the function returns the value of the threshold using threshold_var. If not supplied, the function returns the unchanged value, which can be compared against the version which uses threshold_var.

t_max

The longest time to run project to find steady state.

Value

A data frame with columns F and yield.

See also

Examples

if (FALSE) {
params <- newMultispeciesParams(NS_species_params_gears, inter)
y <- getYieldVsF(params, "Cod", F_max = 1, no_steps = 5)
}