Controlling the species parameters not included in other controls

Adds controls for the parameters that did not fit into other categories.

Usage

otherControl(input, output, session, params, params_old, flags, ...)

otherControlUI(p, input)

Arguments

input

Reactive holding the inputs

output

Reactive holding the outputs

session

Shiny session

params

Reactive value holding updated MizerParams object

params_old

Reactive value holding non-updated MizerParams object

flags

Environment holding flags to skip certain observers

...

Unused

p

The MizerParams object currently being tuned.

Details

The parameters included are:

conversion efficiency alpha

metabolic rate coefficient ks and exponent p

activity rate coefficient k

external mortality at maturity size

If the external mortality at maturity size is changed, then the external mortality at all other sizes is scaled by the same factor.

If the metabolism exponent p is changed, this also changes ks so that metabolic rate at maturity stays the same.