Rescale System

The abundances in mizer and some rates depend on the size of the area to which they refer. So they could be given per square meter or per square kilometer or for an entire study area or any other choice of yours. This function allows you to change the size by automatically changing the abundances and rates accordingly.

rescaleSystem(params, factor)

Arguments

params
A mizer params object

factor
The factor by which the size is rescaled with respect to which the abundances are given.

Value

An object of type MizerParams

Details

If you rescale the system by a factor \(c\) then this function makes the following rescalings in the params object:

The initial abundances initial_n, initial_n_pp and initial_n_other are rescaled by \(c\).

The search volume is rescaled by \(1/c\).

The resource carrying capacity is rescaled by \(c\)

The maximum reproduction rate \(R_{max}\), if used, is rescaled by \(c\).

The effect of this is that the dynamics of the rescaled system are identical to those of the unscaled system, in the sense that it does not matter whether one first calls rescaleSystem and then runs a simulation with project or whether one first runs a simulation and then rescales the resulting abundances.

Note that if you use non-standard resource dynamics or other components then you may need to rescale additional parameters that appear in those dynamics.